+++ src/Enums/DiscountType.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Subtain\LaravelPayments\Enums;
+
+enum DiscountType: string
+{
+    case PERCENTAGE = 'percentage';
+    case FIXED = 'fixed';
+}